pruebaFlinkapp/requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
psycopg2_binary==2.9.3
tzdata == 2021.5
